requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
requests==2.31.0
cachetools==5.3.3
tenacity==8.2.3
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1


